Referral advice for weight for height

This function traverses the decision tree of the "JGZ-Richtlijn overgewicht 2012" and "JGZ-Richtlijn Ondergewicht 2019"

calculate_advice_wgt(
  sex = NA_character_,
  ga = NA,
  etn = NA,
  dom0 = NA_integer_,
  y0 = NA,
  dom1 = NA_integer_,
  y1 = NA,
  hgt1 = NA,
  hgt0 = NA,
  test_gain = TRUE,
  verbose = FALSE
)

Arguments

sex

Character, either "male" or "female"

ga

Gestational age, completed weeks (Integer or character)

etn

Etnicity, one of "NL" (dutch), "TU" (turkish), "MA" (moroccan) or "HS" (hindustani). The default is "NL". Only used for target height.

dom0

Age at previous measurement (days)

y0

Height at previous measurement (cm)

dom1

Age at last measurement (days)

y1

Height at last measurement (cm)

hgt1

Height at last measurement (cm)

hgt0

Height at previous measurement (cm)

test_gain

Logical. Should the increase or decrease in Z-scores be tested? The default is TRUE.

verbose

Set to TRUE to obtain warnings on reference finding.

Value

calculate_advice_wgt() returns an integer, the msgcode

Details

The decision tree assesses both single and paired measurements. The last observations (y1) is generally taken as the last measurement, whereas y0 can be one of the previous measurements. For more than two measurements, there are many pairs possible, and these pairs need not be consecutive. The y0 measurement needs to be defined by the user, and is informally taken as an earlier measurement that maximizes the referral probability.

Author

Arjan Huizing, Stef van Buuren 2020

Examples

msg(calculate_advice_wgt())
#> [1] "Voer het geslacht in."
msgcode <- calculate_advice_wgt(sex = "male", dom1 = 1441, y1 = 25, hgt1 = 120)
msg(msgcode)
#> [1] "Het advies kan niet worden bepaald. Voer het gewicht bij de eerdere meting in."