Table of Dutch birth weight references

Table of Dutch birth weight references

Table of Dutch birth length references

Table of Dutch nl5 references

Table of return message and return code

Table of BMI cut-off points (Cole, 2000 (overweight), 2007 (thinness), De Wilde)

Example of an object of class target created by thge bdsreader package

Data frame of BDS elements and the window of time in which they should be performed

Usage

ref.nl5defSGAgewicht

ref.nl5defSGAlengte

nl5.hgt

messages

bmi_table

target

reminder_table

Format

An object of class data.frame with 36 rows and 8 columns.

An object of class data.frame with 36 rows and 8 columns.

An object of class data.frame with 706 rows and 7 columns.

An object of class data.frame with 101 rows and 2 columns.

An object of class data.frame with 132 rows and 6 columns.

An object of class list of length 2.

An object of class data.frame with 37 rows and 4 columns.