Extract measurements that can function as date zero

This function scans the tgt object, and finds the observation pairs used by the calculate_advice_xxx() functions.

calculate_screening_doms(tgt, ynames = c("hgt", "wgt", "hdc"), na.omit = TRUE)

Arguments

tgt

A list with elements psn and xyz

ynames

Character vector identifying the measures to be screened. By default, ynames = c("hgt", "wgt", "hdc").

na.omit

A logical indicating whether records with a missing x (age) or y (yname) should be removed. Defaults to TRUE.

Value

A list with length(ynames) elements. Each list element is another list with elements dom0 (back-calculated dates of measurement, vector, reverse time), age0 (decimal age), y0 (measurement), z0 (Z-score equivalent) and

for weight - h0, as well as similar quantities dom1, age1, y1, z1 and h1 for the upper part of the pair.

Details

This function implements the "x1_last" strategy for choosing pairs. This strategy selects the last observation as x1 and forms pairs with every earlier observation.

Note

Internal function. Not to be called directly.

Examples

growthscreener:::calculate_screening_doms(target)
#> $hgt
#> $hgt$dom0
#> [1] 562 457 373 275 177  86  58  37   0
#> 
#> $hgt$dom1
#> [1] 745
#> 
#> $hgt$age0
#> [1] 1.5387 1.2512 1.0212 0.7529 0.4846 0.2355 0.1588 0.1013 0.0000
#> 
#> $hgt$age1
#> [1] 2.0397
#> 
#> $hgt$y0
#> [1] 84.0 80.0 75.0 71.5 65.5 59.5 56.0 53.5 48.0
#> 
#> $hgt$y1
#> [1] 90
#> 
#> $hgt$z0
#> [1]  0.527  0.421 -0.180  0.131 -0.259  0.163 -0.261 -0.499 -1.515
#> 
#> $hgt$z1
#> [1] 0.67
#> 
#> $hgt$h0
#> [1] NA NA NA NA NA NA NA NA NA
#> 
#> $hgt$h1
#> [1] NA
#> 
#> 
#> $wgt
#> $wgt$dom0
#> [1] 562 457 373 275 177  86  58  37   0
#> 
#> $wgt$dom1
#> [1] 745
#> 
#> $wgt$age0
#> [1] 1.5387 1.2512 1.0212 0.7529 0.4846 0.2355 0.1588 0.1013 0.0000
#> 
#> $wgt$age1
#> [1] 2.0397
#> 
#> $wgt$y0
#> [1] 12.80 11.90 10.95  9.65  8.24  5.90  5.00  4.18  2.95
#> 
#> $wgt$y1
#> [1] 13.9
#> 
#> $wgt$z0
#> [1]  1.228  1.247  1.164  1.052  1.173  0.717  0.401 -0.162 -1.055
#> 
#> $wgt$z1
#> [1] 0.989
#> 
#> $wgt$h0
#> [1] 84.0 80.0 75.0 71.5 65.5 59.5 56.0 53.5 48.0
#> 
#> $wgt$h1
#> [1] 90
#> 
#> 
#> $hdc
#> $hdc$dom0
#> [1] 562 457 373 275 177  86  58  37
#> 
#> $hdc$dom1
#> [1] 745
#> 
#> $hdc$age0
#> [1] 1.5387 1.2512 1.0212 0.7529 0.4846 0.2355 0.1588 0.1013
#> 
#> $hdc$age1
#> [1] 2.0397
#> 
#> $hdc$y0
#> [1] 49.2 48.7 47.8 46.6 44.1 40.5 39.0 37.6
#> 
#> $hdc$y1
#> [1] 50
#> 
#> $hdc$z0
#> [1] 1.246 1.373 1.307 1.418 1.021 0.696 0.605 0.418
#> 
#> $hdc$z1
#> [1] 1.224
#> 
#> $hdc$h0
#> [1] NA NA NA NA NA NA NA NA
#> 
#> $hdc$h1
#> [1] NA
#> 
#>