Screen growth curves according to JGZ guidelines

Screen growth curves according to JGZ guidelines

screen_curves_ind(
  ind,
  ynames = c("hgt", "wgt", "hdc"),
  na.omit = TRUE,
  recalculate_z = FALSE
)

Arguments

ind

A tibble with a person attribute. Support for minihealth::individual object is deprecated.

ynames

Character vector identifying the measures to be screened. By default, ynames = c("hgt", "wgt", "hdc").

na.omit

A logical indicating whether records with a missing x (age) or y (yname) should be removed. Defaults to TRUE.

recalculate_z

A logical indicating whether Z-scores should be recalculated. Currently not functional. Z-scores are always recalculated.

Value

A data frame with the following columns

Categorie

The category of the screening guidelines: 1000 = hgt, 2000 = wgt, 3000 = hdc

CategorieOmschrijving

A string indicating the measure

Code

Integer message code

CodeOmschrijving

A string with the message

Versie

Version of growthscreener package

Leeftijd0

First date of the evaluation pair

Leeftijd1

Second date of the evaluation pair

Examples

screen_curves_ind(target)
#>    Categorie CategorieOmschrijving Code
#> 1       1000                Lengte 1031
#> 2       1000                Lengte 1031
#> 3       1000                Lengte 1031
#> 4       1000                Lengte 1031
#> 5       1000                Lengte 1031
#> 6       1000                Lengte 1031
#> 7       1000                Lengte 1031
#> 8       1000                Lengte 1031
#> 9       1000                Lengte 1025
#> 10      2000               Gewicht 2031
#> 11      2000               Gewicht 2031
#> 12      2000               Gewicht 2031
#> 13      2000               Gewicht 2031
#> 14      2000               Gewicht 2075
#> 15      2000               Gewicht 2031
#> 16      2000               Gewicht 2031
#> 17      2000               Gewicht 2031
#> 18      2000               Gewicht 2031
#> 19      3000           Hoofdomtrek 3021
#> 20      3000           Hoofdomtrek 3021
#> 21      3000           Hoofdomtrek 3021
#> 22      3000           Hoofdomtrek 3021
#> 23      3000           Hoofdomtrek 3021
#> 24      3000           Hoofdomtrek 3021
#> 25      3000           Hoofdomtrek 3021
#> 26      3000           Hoofdomtrek 3021
#>                                                                                                                                                                                                                                                                                                                                                          CodeOmschrijving
#> 1                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 2                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 3                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 4                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 5                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 6                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 7                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 8                                                                                                                                                                                                                                           Het advies volgens de JGZ-richtlijn lengtegroei is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 9                                                                                                                                                                                                                                                                                                                          Kan de SDS van de eerdere meting niet bepalen.
#> 10                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 11                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 12                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 13                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 14 Het advies volgens de JGZ-richtlijn ondergewicht is als volgt: Sterke gewichtsafname (-1 SD), advies: Is er sprake van een afwijkende voedingstoestand en/of klachten of symptomen die kunnen wijzen op onderliggende ziekte of problemen? Indien ja, Verwijzen naar kinderarts. Indien nee, dan is er in principe geen verwijzing nodig. Naar eigen inzicht handelen.
#> 15                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 16                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 17                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 18                                                                                                                                                                                                                                          Het advies volgens de JGZ-richtlijn overgewicht is als volgt: In principe geen verwijzing nodig, naar eigen inzicht handelen.
#> 19                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 20                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 21                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 22                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 23                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 24                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 25                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#> 26                                                                                                                                                                                                                                                                                                          De richtlijn hoofdomtrek is bedoeld voor kinderen tot 1 jaar.
#>    Versie Leeftijd0 Leeftijd1
#> 1  1.11.0  19900806  19910205
#> 2  1.11.0  19900423  19910205
#> 3  1.11.0  19900129  19910205
#> 4  1.11.0  19891023  19910205
#> 5  1.11.0  19890717  19910205
#> 6  1.11.0  19890417  19910205
#> 7  1.11.0  19890320  19910205
#> 8  1.11.0  19890227  19910205
#> 9  1.11.0  19890121  19910205
#> 10 1.11.0  19900806  19910205
#> 11 1.11.0  19900423  19910205
#> 12 1.11.0  19900129  19910205
#> 13 1.11.0  19891023  19910205
#> 14 1.11.0  19890717  19910205
#> 15 1.11.0  19890417  19910205
#> 16 1.11.0  19890320  19910205
#> 17 1.11.0  19890227  19910205
#> 18 1.11.0  19890121  19910205
#> 19 1.11.0  19900806  19910205
#> 20 1.11.0  19900423  19910205
#> 21 1.11.0  19900129  19910205
#> 22 1.11.0  19891023  19910205
#> 23 1.11.0  19890717  19910205
#> 24 1.11.0  19890417  19910205
#> 25 1.11.0  19890320  19910205
#> 26 1.11.0  19890227  19910205